Committing for the 3rd time as sourceforge seems to be throwing a wobbly...
Rather then use a batch script under windows and a shell script under *nix to start the CCP1GUI, it makes more sense to use a python script as this way we have more control and a more unified interface on all platforms.

It also means that we can pop up dialog boxes with relevant information if modules aren't found - something that is vital under Windows as otherwise the script just spits output to the DOS shell which the user never sees.

Now, we can pop up dialogs to tell the user what they are missing and where they can find it.

Currently we advise the users to get the vtkpython stuff from Christoph Gohlke's binaries, as Enthought are no longer releasing their python distribution for free.

This file is executable, so under any unix system where python is in /usr/bin/python, it will just execute as a shell script that will invoke python.