Change so that the gui will ask for the name of the input file to write when "Write Input" is selected in the calculation editor.
There was a get_name and set_name method in the calculation class  that I'd not noticed before so I've changed things to use this to make the interface more uniform across the different codes.

I've had to hack the TextInput tool to make it use the get/set_name method when it's used for the name.

This fixes issue 2992808 on the sourceforge bug tracker.